wip: Fix frame children coords when dragging position stats coords
