+++ packages/excalidraw/components/Stats/Angle.tsx
@@ -12,9 +12,8 @@ import type Scene from "@excalidraw/element/Scene";
 import { angleIcon } from "../icons";
 
 import DragInput from "./DragInput";
-import { getStepSizedValue, isPropertyEditable, updateBindings } from "./utils";
+import { DragInputCallbackType, getStepSizedValue, isPropertyEditable, updateBindings } from "./utils";
 
-import type { DragInputCallbackType } from "./DragInput";
 import type { AppState } from "../../types";
 
 interface AngleProps {
