+++ packages/excalidraw/components/Stats/Dimension.tsx
@@ -13,9 +13,8 @@ import type { ExcalidrawElement } from "@excalidraw/element/types";
 import type Scene from "@excalidraw/element/Scene";
 
 import DragInput from "./DragInput";
-import { getStepSizedValue, isPropertyEditable } from "./utils";
+import { DragInputCallbackType, getStepSizedValue, isPropertyEditable } from "./utils";
 
-import type { DragInputCallbackType } from "./DragInput";
 import type { AppState } from "../../types";
 
 interface DimensionDragInputProps {
