+++ packages/excalidraw/components/Stats/FontSize.tsx
@@ -17,9 +17,8 @@ import type Scene from "@excalidraw/element/Scene";
 import { fontSizeIcon } from "../icons";
 
 import StatsDragInput from "./DragInput";
-import { getStepSizedValue } from "./utils";
+import { DragInputCallbackType, getStepSizedValue } from "./utils";
 
-import type { DragInputCallbackType } from "./DragInput";
 import type { AppState } from "../../types";
 
 interface FontSizeProps {
