+++ packages/excalidraw/components/Stats/MultiDimension.tsx
@@ -26,9 +26,8 @@ import type Scene from "@excalidraw/element/Scene";
 
 import DragInput from "./DragInput";
 import { getAtomicUnits, getStepSizedValue, isPropertyEditable } from "./utils";
-import { getElementsInAtomicUnit } from "./utils";
+import { DragInputCallbackType, getElementsInAtomicUnit } from "./utils";
 
-import type { DragInputCallbackType } from "./DragInput";
 import type { AtomicUnit } from "./utils";
 import type { AppState } from "../../types";
