+++ packages/excalidraw/components/Stats/utils.ts
@@ -5,7 +5,9 @@ import {
   updateBoundElements,
 } from "@excalidraw/element/binding";
 import { getBoundTextElement } from "@excalidraw/element/textElement";
+import { getCommonBounds } from "@excalidraw/element/bounds";
 import {
+  isFrameChildElement,
   isFrameLikeElement,
   isLinearElement,
   isTextElement,
@@ -37,6 +39,23 @@ export type StatsInputProperty =
   | "angle"
   | "fontSize"
   | "gridStep";
+
+export type DragInputCallbackType<
+  P extends StatsInputProperty,
+  E = ExcalidrawElement,
+> = (props: {
+  accumulatedChange: number;
+  instantChange: number;
+  originalElements: readonly E[];
+  originalElementsMap: ElementsMap;
+  shouldKeepAspectRatio: boolean;
+  shouldChangeByStepSize: boolean;
+  scene: Scene;
+  nextValue?: number;
+  property: P;
+  originalAppState: AppState;
+  setInputValue: (value: number) => void;
+}) => void;
 
 export const SMALLEST_DELTA = 0.01;
 
@@ -73,9 +92,9 @@ export const getElementsInAtomicUnit = (
       latest: elementsMap.get(id),
     }))
     .filter((el) => el.original !== undefined && el.latest !== undefined) as {
-    original: NonDeletedExcalidrawElement;
-    latest: NonDeletedExcalidrawElement;
-  }[];
+      original: NonDeletedExcalidrawElement;
+      latest: NonDeletedExcalidrawElement;
+    }[];
 };
 
 export const newOrigin = (
@@ -174,6 +193,89 @@ export const moveElement = (
   }
 };
 
+export const moveElements = (
+  property: "x" | "y",
+  changeInTopX: number,
+  changeInTopY: number,
+  originalElements: readonly ExcalidrawElement[],
+  originalElementsMap: ElementsMap,
+  scene: Scene,
+) => {
+  for (let i = 0; i < originalElements.length; i++) {
+    const origElement = originalElements[i];
+
+    const [cx, cy] = [
+      origElement.x + origElement.width / 2,
+      origElement.y + origElement.height / 2,
+    ];
+    const [topLeftX, topLeftY] = pointRotateRads(
+      pointFrom(origElement.x, origElement.y),
+      pointFrom(cx, cy),
+      origElement.angle,
+    );
+
+    const newTopLeftX =
+      property === "x" ? Math.round(topLeftX + changeInTopX) : topLeftX;
+
+    const newTopLeftY =
+      property === "y" ? Math.round(topLeftY + changeInTopY) : topLeftY;
+
+    moveElement(
+      newTopLeftX,
+      newTopLeftY,
+      origElement,
+      scene,
+      originalElementsMap,
+      false,
+    );
+  }
+};
+
+export const moveGroup = (
+  nextX: number,
+  nextY: number,
+  originalElements: ExcalidrawElement[],
+  originalElementsMap: ElementsMap,
+  scene: Scene,
+) => {
+  const elementsMap = scene.getNonDeletedElementsMap();
+  const [x1, y1, ,] = getCommonBounds(originalElements);
+  const offsetX = nextX - x1;
+  const offsetY = nextY - y1;
+
+  for (let i = 0; i < originalElements.length; i++) {
+    const origElement = originalElements[i];
+
+    const latestElement = elementsMap.get(origElement.id);
+    if (!latestElement) {
+      continue;
+    }
+
+    // bound texts are moved with their containers
+    if (!isTextElement(latestElement) || !latestElement.containerId) {
+      const [cx, cy] = [
+        latestElement.x + latestElement.width / 2,
+        latestElement.y + latestElement.height / 2,
+      ];
+
+      const [topLeftX, topLeftY] = pointRotateRads(
+        pointFrom(latestElement.x, latestElement.y),
+        pointFrom(cx, cy),
+        latestElement.angle,
+      );
+
+      moveElement(
+        topLeftX + offsetX,
+        topLeftY + offsetY,
+        origElement,
+        scene,
+        originalElementsMap,
+        false,
+      );
+    }
+  }
+};
+
 export const getAtomicUnits = (
   targetElements: readonly ExcalidrawElement[],
   appState: AppState,
@@ -210,3 +312,116 @@ export const updateBindings = (
     updateBoundElements(latestElement, scene, options);
   }
 };
+
+export const handlePositionChange: DragInputCallbackType<
+  "x" | "y"
+> = ({
+  accumulatedChange,
+  originalElements,
+  originalElementsMap,
+  shouldChangeByStepSize,
+  nextValue,
+  property,
+  scene,
+  originalAppState,
+}) => {
+    const STEP_SIZE = 10;
+    const elementsMap = scene.getNonDeletedElementsMap();
+
+    if (nextValue !== undefined) {
+      for (const atomicUnit of getAtomicUnits(
+        originalElements,
+        originalAppState,
+      )) {
+        const elementsInUnit = getElementsInAtomicUnit(
+          atomicUnit,
+          elementsMap,
+          originalElementsMap,
+        );
+
+        if (elementsInUnit.length > 1) {
+          const [x1, y1, ,] = getCommonBounds(
+            elementsInUnit.map((el) => el.latest!),
+          );
+          const newTopLeftX = property === "x" ? nextValue : x1;
+          const newTopLeftY = property === "y" ? nextValue : y1;
+
+          moveGroup(
+            newTopLeftX,
+            newTopLeftY,
+            elementsInUnit.map((el) => el.original),
+            originalElementsMap,
+            scene,
+          );
+        } else {
+          const origElement = elementsInUnit[0]?.original;
+          const latestElement = elementsInUnit[0]?.latest;
+          if (
+            origElement &&
+            latestElement &&
+            isPropertyEditable(latestElement, property)
+          ) {
+            const [cx, cy] = [
+              origElement.x + origElement.width / 2,
+              origElement.y + origElement.height / 2,
+            ];
+            const [topLeftX, topLeftY] = pointRotateRads(
+              pointFrom(origElement.x, origElement.y),
+              pointFrom(cx, cy),
+              origElement.angle,
+            );
+
+            if (isFrameChildElement(origElement)) {
+              const childNewTopLeftX = property === "x" ? nextValue + Math.abs(topLeftX) : topLeftX;
+              const childNewTopLeftY = property === "y" ? nextValue + Math.abs(topLeftY) : topLeftY;
+
+              moveElement(
+                childNewTopLeftX,
+                childNewTopLeftY,
+                origElement,
+                scene,
+                originalElementsMap,
+                false,
+              );
+
+              scene.triggerUpdate();
+              return;
+            } else {
+              const newTopLeftX = property === "x" ? nextValue : topLeftX;
+              const newTopLeftY = property === "y" ? nextValue : topLeftY;
+
+              moveElement(
+                newTopLeftX,
+                newTopLeftY,
+                origElement,
+                scene,
+                originalElementsMap,
+                false,
+              );
+            }
+          }
+        }
+      }
+
+      scene.triggerUpdate();
+      return;
+    }
+
+    const change = shouldChangeByStepSize
+      ? getStepSizedValue(accumulatedChange, STEP_SIZE)
+      : accumulatedChange;
+
+    const changeInTopX = property === "x" ? change : 0;
+    const changeInTopY = property === "y" ? change : 0;
+
+    moveElements(
+      property,
+      changeInTopX,
+      changeInTopY,
+      originalElements,
+      originalElementsMap,
+      scene,
+    );
+
+    scene.triggerUpdate();
+  };
